MMR Pruning and Cleanup (#47)
* Prevent MMR data files from getting out of sync with the DB

* Require kernel uniqueness within blocks

* Cleanup Logger

* Removing LoggerAPI::Shutdown references

* Removing ThreadManager

* Minor cleanup

* Indexing leaves by leaf_idx in LeafDB

* Efficient MMR node iterator by height

* BuildCompactBitSet tests

* Test for DiffCompactBitSet

* Adding PruneList

* Cleanup

* BitSet class

* Cleanup

* Schnorr signature cache and cleanup

* Cmake project cleanup

* Stubbing IsMine wallet function

+++ src/libmw/libmw_wallet.cpp
@@ -53,7 +53,7 @@ MWEXPORT libmw::MWEBAddress GetAddress(const libmw::IWallet::Ptr& pWallet, const
 
 MWEXPORT bool IsOwnAddress(const libmw::IWallet::Ptr& pWallet, const libmw::MWEBAddress& address)
 {
-    for (size_t i = 0; i < 100; i++) {
+    for (uint32_t i = 0; i < 100; i++) {
         if (GetAddress(pWallet, i) == address) {
             return true;
         }
